code in index.js seperated to their own js files for each react component
